+++ src/SortedList.h
@@ -11,7 +11,7 @@ struct node_sorted_list
 typedef struct info_sorted_list {
 	struct node_sorted_list *header;
 	struct node_sorted_list *tail;
-	unsigned n_elements;
+	unsigned long n_elements;
 	int (*f_compare)(void *, void *);
 } *sorted_list;
 
@@ -31,7 +31,7 @@ int isEmpty_sorted_list(sorted_list l);
    ERROR: If list is not initialized
 */
 
-unsigned size_sorted_list(sorted_list l);
+unsigned long size_sorted_list(sorted_list l);
 /* NEEDS: A list already initialized
    RETURNS: The number of elements in the list
    ERROR: If list is not initialized
